@@ -4,18 +4,20 @@
   "description": "Node app built on express for Google Maps traffic alerts.",
   "main": "src/server.js",
   "scripts": {
-    "start": "node src/server.js",
+    "start": "node -r esm src/server.js",
     "dev": "nodemon --watch src src/server.js"
   },
   "dependencies": {
     "@google/maps": "^1.0.2",
+    "axios": "^0.19.0",
     "dotenv": "^8.2.0",
     "ejs": "^2.6.1",
+    "esm": "^3.2.25",
     "express": "^4.16.4",
     "nodemon": "^2.0.1"
   },
   "engines": {
-    "node": "8.x"
+    "node": ">=12.0.0"
   },
   "repository": {
     "url": "https://ifttt.com"
